WIP: Storage encryption
This is a work in progress for the storage encryption.
Currently, the symmetric key, which we will retrieve from the
keyshare server, is hardcoded. This patch includes some tests
and an update function to convert the unencrypted storage to encrypted.
We should also think about using the key from the keyshare
server as a key encryption key and not directly. This allows replacing
the encryption key.